Release v0.4.0

+++ test/playground_integration/catalog_integration_test.go
@@ -36,6 +36,9 @@ func cleanupDatasets(t *testing.T) {
 	assert.Emptyf(t, err, "Error retrieving the datasets: %s", err)
 
 	for _, item := range result {
+		if item.Name=="main" || item.Name=="metrics" {
+			continue
+		}
 		err = client.CatalogService.DeleteDataset(item.ID)
 		assert.Emptyf(t, err, "Error deleting dataset: %s", err)
 	}
